+++ src/data/selector.ts
@@ -1,0 +1,16 @@
+export type Selector = EntitySelector | DeviceSelector;
+
+export interface EntitySelector {
+  entity: {
+    integration?: string;
+    domain?: string;
+  };
+}
+
+export interface DeviceSelector {
+  device: {
+    integration?: string;
+    manufacturer?: string;
+    model?: string;
+  };
+}
